agent/requirements.txt
# Gold Evaluator Agent Dependencies
# Generated for gold_evaluator.py and related modules

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# OpenTelemetry and Phoenix tracing
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-http>=1.21.0

# Encryption
cryptography>=41.0.0

# Standard library modules (no additional packages needed)
# - json
# - os
# - sys
# - uuid
# - datetime
# - typing
# - hashlib
